Refactored BeerDTO and BeerService to use Double values for AlcoholByVolume

+++ BeerOverflow/BeerOverflow.Services/DTO/Contracts/IBeerDTO.cs
@@ -11,6 +11,6 @@ namespace BeerOverflow.Services.DTO.Contracts
         string Name { get; set; }
         BeerType BeerType { get; set; }
         string Brewery { get; set; }
-        int AlcoholByVolume { get; set; }
+        double AlcoholByVolume { get; set; }
     }
 }
